line: Improve algorithm
This works far better, there's still plenty of improvements to be done.
In particular, the search after the line is lost should make used of the
IMU to determine how far to turn instead of using time.

I could probably squeeze out a bit more speed too, it will mean tweaking
the factors for deceleration and turn-speed.