bounce: Try and avoid jumps.
If the ground's not flat, when going up a bump it looks like the horizon
is closer than it really is. Try and mitigate that by waiting for 5
frames before we decide it really is time to turn, droppping the speed
aggressively in the meantime.

Not sure whether to use this on the day - depends how the real course
looks.